stream: allocate send RangeBuf with consistent sizes
This changes how RangeBuf objects are allocated when sending data.

Currently a new RangeBuf is allocated for each application write (of the
size of the write itself).

This change makes it possible to allocate RangeBufs of a pre-defined
capacity, whether the application data can fill it or not:

 * If the application data is bigger than the RangeBuf capacity, new
   RangeBuf objects are allocated (with the same fixed capacity) until
   all the application data has been stored.

 * If the application data is smaller than the RangeBuf capacity, the
   RangeBuf is stored with the unused capacity at the back of the
   SendBuf. Following writes will then try to fill the trailing RangeBuf
   if it has any spare capacity.

The use of Arc<RwLock<>> for the RangeBuf's internal buffer makes
accessing the data somewhat more complicated as we can't slice the
Vec<u8> directly anymore, which is why the `with()` method was added.

However this added complexity is necessary to make sure Connection can
implement the Send trait, and due to the fact that Arc requires its
inner type to be both Sync and Send, we need to use RwLock (instead of,
say, RefCell) to satisfy the type constraints.