ensure there is enough space for the minimum payload length
In order for packet encryption to work properly, the payload needs to be
at least 4 bytes. When that is not the case, PADDING frames are added to
reach the minimum length.

It might happen however that there is no space left in the packet to add
this padding. Before commit 3d62f9a360e9b889e82cbf0a37b1932ac421c479 we
would simply add the padding anyway, however in that commit, the
relevant code was changed to use push_frame_to_pkt!() to add the frame
to the packet, which will prevent adding more bytes than allowed. This
in turn lead to the padding not being added in some cases (e.g. when the
server is blocked by the anti-amplification limit).

Instead of forcing the frame to be added even if there is no space,
simply return early on if we can't write enough bytes.