h3: ensure Event::Finished is raised for empty STREAM frames with fin flag
If the server is finishing the stream with sending an empty frame with
the FIN bit set after the client has recv'd the body of the response,
the subsequent poll is returning Error::Done instead of raising
Event::Finished. If there is no more data to be received, the client
might not call poll again. Thus, make sure to raise the event properly
in such cases.

Fixes #740.