properly handle incoming STOP_SENDING frames
This implements resetting a stream's send-side, as well as sending a
RESET_STREAM frame, upon receipt of STOP_SENDING from the peer.

By resetting a stream, all of its already buffered data is dropped, and
no more writes are permitted.

In order to signal to the application that a stream was "stopped" and
reset, the stream is marked as writable upon receipt of the STOP_SENDING
frame, but when calling `stream_send()` as well as `stream_capacity()` a
new error, `StreamStopped` will be returned.

Returning the error from the `stream_capacity()` is required to make
sure the application is notified in cases where the method is called to
check whether any capacity is available before calling `stream_send()`
(like we do in the HTTP/3 code), and the target stream is blocked (so
`stream_send()` isn't called at all).

The `StreamStopped` error carries the raw error code sent as part of the
STOP_SENDING frame, so that an application can treat the error case
differently depending on the meaning of the code.

Currently the error is turned into a `TransportError` at the HTTP/3
layer. In the future we can better translate this into an appropriate
HTTP/3 error.

Fixes #585.