h3: limit the size of SETTINGS frame that quiche is willing to parse
Previously, quiche would parse SETTINGS frames with any valid size.
There are only a few settings parameters defined in the core HTTP/3
specification, and although we should accomodate extensions or GREASE,
it seems unlikely that we'll receive legitimate frames that are large.

This change adds an implementation limit for the size of SETTINGS
frames. Frames larger than the limit will trigger a connection error.