limit send capacity over multiple stream_send() calls
A connection send capacity is limited by the size of the flow and
congestion windows. However, cwnd is not reduced between stream_send()
calls, which might cause more data than necessary to be buffered.

Instead of using the current available cwnd size for each call, buffer
the value and reuse it over a single "flight".