stream: process incoming data when draining
When a stream enters the draining state, besides not storing the
received data, it also stops updating its internal state, like the
stream's maximum offset, which affects the flow control calculation.

For example, if the stream is reset by the peer, this could trigger a
spurious flow control violation error, causing the connection to be
terminated.