h3: emit GOAWAY from client
Previously, if `send_goaway()` was called on a client connection, it
would return without emitting a frame. This was justified by the lack
of server push support in quiche. However, it also prevented client
GOAWAY when the endpoints didn't use push, which is unfortunate.

This change simply always populates the emitted GOAWAY frame with an
ID of 0, which is valid and consistent with our server push support
status.