add PADDING frame to fill up the datagram with Initial packet
For Initial packets we need to make the UDP datagram at least 1200
bytes following the transport spec. However when the last packet
in the datagram is Short and the datagram is not fully filled, we try
to do null padding at the end which causes a decryption failure on
the server because Short packet doesn't have an explicit length.

To avoid this issue, when Short packet is added in the same
datagram with Initial packet, add a PADDING frame to Short packet
to fill up the datagram size. If Short packet is not added,
we still add zero padding at the end of packet.