qlog: add means to provide external event time
Previously, the `add_event()` method populated the qlog event_time
field by calling `std::time::Instant::elapsed()`, which causes an in
direct call for `std::time::Instand::now()`. This incurs additional
system calls that are wasteful, especially if events are logged in
rapid succession or the time is available elsewhere.

This change introduces the `add_event_with_instant()` method that
allows callers to provide their own Instant that they wish to use
when logging an event.