cc: add send_quantum API
For higher performance sending multiple packets is a common
practice, such as sendmmsg() or linux sendmsg() GSO. To
help with that, send_quantum() is added and it can be modified
in the congestion controller. It's a hint for the application,
it may or may not use the value for controlling the batch
size of packets.

For default value, initcwnd is used(RFC9002 section 7.7).

Also, conn.send_quantum() and C API (quiche_conn_send_quantum())
is provided for the application.