pacing: no pacing can cause out-of-order
Especially for ACK packet (or other packet with no pacing
condition) which has no payload size, it need to be
sent immediately. However if next scheduled time is already
in the future, sending ACK immediately can cause out-or-order
delivery and result in lost packet and retransmit.

To prevent such case, if `last_packet_scheduled_time` is
already in the future when no pacing used, don't update
the timestamp. Otherwise the current time(now) will be used.